+++ src/nodes/scroll/elements/WindowsClassicArrow.cpp
@@ -5,7 +5,7 @@ using namespace geode::prelude;
 using namespace alpha::prelude;
 
 struct WindowsClassicArrow::Impl final {
-    cocos2d::extension::CCScale9Sprite* m_background;
+    geode::NineSlice* m_background;
     cocos2d::CCSprite* m_arrow;
     bool m_holding = false;
 };
@@ -30,7 +30,7 @@ bool WindowsClassicArrow::init() {
     m_impl->m_arrow->setColor({0, 0, 0});
     m_impl->m_arrow->setZOrder(1);
 
-    m_impl->m_background = CCScale9Sprite::create("WindowsClassic.png"_spr);
+    m_impl->m_background = geode::NineSlice::create("WindowsClassic.png"_spr);
     m_impl->m_background->setScale(0.5f);
 
     setContentHeight(10.f);