+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.0.1-beta.2
+- Geode Alpha 2 support
+- Switch to geode::NineSlice
+- Disable touch blocker by default (touch prio is funky), it is easier to handle that on your own as a dev
+- Bug Fixes
+
 # 1.0.1-beta.1
 - 2.2081 Support
 - Fix minor bug when dragging with anything other than left click